chore: move dependency installation to `init` command
