+++ cmd/binaries/binaries.go
@@ -1,18 +1,19 @@
 package binaries
 
-import (
-	"github.com/spf13/cobra"
-
-	"github.com/dymensionxyz/roller/cmd/binaries/install"
-)
-
-func Cmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "binaries",
-		Short: "Commands to manage roller dependencies",
-	}
-
-	cmd.AddCommand(install.Cmd())
-
-	return cmd
-}
+//
+// import (
+// 	"github.com/spf13/cobra"
+//
+// 	"github.com/dymensionxyz/roller/cmd/binaries/install"
+// )
+//
+// func Cmd() *cobra.Command {
+// 	cmd := &cobra.Command{
+// 		Use:   "binaries",
+// 		Short: "Commands to manage roller dependencies",
+// 	}
+//
+// 	cmd.AddCommand(install.Cmd())
+//
+// 	return cmd
+// }
