+++ cmd/root.go
@@ -3,9 +3,6 @@ package cmd
 import (
 	"os"
 
-	"github.com/spf13/cobra"
-
-	"github.com/dymensionxyz/roller/cmd/binaries"
 	blockexplorer "github.com/dymensionxyz/roller/cmd/block-explorer"
 	da_light_client "github.com/dymensionxyz/roller/cmd/da-light-client"
 	"github.com/dymensionxyz/roller/cmd/eibc"
@@ -13,6 +10,7 @@ import (
 	"github.com/dymensionxyz/roller/cmd/rollapp"
 	rollerutils "github.com/dymensionxyz/roller/cmd/utils"
 	"github.com/dymensionxyz/roller/cmd/version"
+	"github.com/spf13/cobra"
 )
 
 var rootCmd = &cobra.Command{
@@ -35,7 +33,7 @@ func init() {
 	// rootCmd.AddCommand(version.Cmd())
 	rootCmd.AddCommand(da_light_client.DALightClientCmd())
 	rootCmd.AddCommand(relayer.Cmd())
-	rootCmd.AddCommand(binaries.Cmd())
+	// rootCmd.AddCommand(binaries.Cmd())
 	// rootCmd.AddCommand(keys.Cmd())
 	// rootCmd.AddCommand(run.Cmd())
 	// rootCmd.AddCommand(services.Cmd())
